Set up custom buttons on app resume
